Move SelectRelatedManager to a common file
